% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sig-ind.R
\name{.calc_sig_ind}
\alias{.calc_sig_ind}
\title{Make a significant indicator}
\usage{
.calc_sig_ind(p_value, format = "html")
}
\arguments{
\item{p_value}{A numeric or numeric vector}

\item{format}{Doesn't do anything right now. Place holder for later. Default
is "html".}
}
\value{
A character or character vector
}
\description{
Adds asterisks next to p-values like \code{summary} does with regression output in
R.
}
