% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-logistic-univ.R
\name{.do_logistic_univ}
\alias{.do_logistic_univ}
\title{Do univariable logistic regression and extract results in nice format.}
\usage{
.do_logistic_univ(
  data,
  formula,
  format = "html",
  conf_level = 0.95,
  exponentiate = TRUE,
  include_last_row = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame or tibble}

\item{formula}{A character string}

\item{format}{Display format in case I need to escape some characters. A
place holder for now in case I need it in the future. Default is "html".}

\item{conf_level}{The confidence level to use for the confidence interval.
Must be strictly greater than 0 and less than 1. Defaults to 0.95, which
corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults to
\code{TRUE}.}

\item{include_last_row}{Adds a row at the end of each set of results to give
some breathing room. Default is \code{TRUE}.}

\item{...}{Additional arguments.}
}
\value{
A tibble or data frame
}
\description{
A helper function to be used in a loop to do univariable regression and give
some nice lookin' results/
}
