% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypval.R
\name{.mypval}
\alias{.mypval}
\title{Formatting p-values}
\usage{
.mypval(pvals, sig_limit = 0.001, digits = 3L, html = FALSE)
}
\arguments{
\item{pvals}{A numeric value or vector of p-values}

\item{sig_limit}{Lower bound for precisions}

\item{digits}{Integer indicating the number of decimal places (round)}

\item{html}{Logical; if \code{TRUE}, uses \code{&lt;} instead of \code{<}.}
}
\description{
A little helper to format p-values in the table.
}
