% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myround.R
\name{.myround}
\alias{.myround}
\title{Do some simple rounding}
\usage{
.myround(x, digits = 1L)
}
\arguments{
\item{x}{A numeric or numeric vector.}

\item{digits}{Integer; number of decimals to round to.}
}
\value{
A character or character vector.
}
\description{
Just a helper function to do some rounding here and there.
}
