% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat-last.R
\name{.repeat_last}
\alias{.repeat_last}
\title{Last observation carried forward}
\usage{
.repeat_last(x, forward = TRUE, maxgap = Inf, na.rm = FALSE)
}
\arguments{
\item{x}{A vector}

\item{forward}{Logical; default is \code{TRUE}. Direction to carry forward. By
specifying forward = \code{FALSE}, you can carry the last observation backward.}

\item{maxgap}{Numeric; By specifying maxgap, you can choose not to bridge
overly long gaps.}

\item{na.rm}{Logical; default is \code{FALSE}. Whether to remove NAs.}
}
\value{
a vector similar to x.
}
\description{
A helper function for making the multivariable regression table
}
\references{
https://stackoverflow.com/questions/7735647/replacing-nas-with-latest-non-na-value
}
