% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purposeful.R
\docType{package}
\name{purposeful-package}
\alias{purposeful}
\alias{purposeful-package}
\title{purposeful: Purposeful selection for regression models}
\description{
Application of the purposeful selection procedure developed by Hosmer and
Lemeshow for selecting predictors in regression models.  The package
provides functions for logistic regression (\code{\link{purposeful_glm}}),
Cox proportional hazards regression (\code{\link{purposeful_cox}}), and
ordinary linear regression (\code{\link{purposeful_lm}}).
}
\examples{
# See ?purposeful_glm, ?purposeful_cox, ?purposeful_lm for full examples.

}
\author{
\strong{Maintainer}: Emile Latour \email{latour@ohsu.edu}

Authors:
\itemize{
  \item Emile Latour \email{latour@ohsu.edu}
}

}
