% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purposeful.R
\docType{package}
\name{purposeful}
\alias{purposeful}
\alias{purposeful-package}
\title{purposeful: Purposeful selection in logistic regression}
\description{
Application of the purposeful selection procedure developed by Homser and
Lemeshow for selecting predictors in logistic regresssion.
}
\examples{
# Example usage:
library(purposeful)

}
